requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via strands-agents-tools
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   aiohttp
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.5
    # via fastmcp
aws-requests-auth==0.4.3
    # via strands-agents-tools
awscrt==0.29.2
    # via botocore
beartype==0.22.8
    # via
    #   py-key-value-aio
    #   py-key-value-shared
beautifulsoup4==4.14.3
    # via markdownify
bedrock-agentcore==1.1.1
    # via tdmcpagentcore (pyproject.toml)
boto3==1.42.5
    # via
    #   bedrock-agentcore
    #   strands-agents
botocore==1.42.5
    # via
    #   tdmcpagentcore (pyproject.toml)
    #   bedrock-agentcore
    #   boto3
    #   s3transfer
    #   strands-agents
    #   strands-agents-tools
cachetools==6.2.2
    # via py-key-value-aio
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
cfgv==3.5.0
    # via pre-commit
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   halo
    #   log-symbols
cryptography==46.0.3
    # via
    #   authlib
    #   pyjwt
cyclopts==4.3.0
    # via fastmcp
dill==0.4.0
    # via strands-agents-tools
diskcache==5.6.3
    # via py-key-value-aio
distlib==0.4.0
    # via virtualenv
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via
    #   cyclopts
    #   strands-agents
docutils==0.22.3
    # via rich-rst
email-validator==2.3.0
    # via pydantic
exceptiongroup==1.3.1
    # via fastmcp
fastmcp==2.13.2
    # via teradata-mcp-server
filelock==3.20.0
    # via virtualenv
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
halo==0.0.31
    # via strands-agents-builder
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   fastmcp
    #   mcp
    #   ollama
httpx-sse==0.4.3
    # via mcp
identify==2.6.15
    # via pre-commit
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via opentelemetry-api
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.25.1
    # via
    #   mcp
    #   strands-agents
jsonschema-path==0.3.4
    # via fastmcp
jsonschema-specifications==2025.9.1
    # via jsonschema
log-symbols==0.0.14
    # via halo
markdown-it-py==4.0.0
    # via rich
markdownify==1.2.2
    # via strands-agents-tools
mcp==1.23.3
    # via
    #   tdmcpagentcore (pyproject.toml)
    #   fastmcp
    #   strands-agents
    #   teradata-mcp-server
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
nodeenv==1.9.1
    # via pre-commit
ollama==0.6.1
    # via strands-agents
openapi-pydantic==0.5.1
    # via fastmcp
opentelemetry-api==1.39.0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-threading
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   strands-agents
opentelemetry-instrumentation==0.60b0
    # via opentelemetry-instrumentation-threading
opentelemetry-instrumentation-threading==0.60b0
    # via strands-agents
opentelemetry-sdk==1.39.0
    # via strands-agents
opentelemetry-semantic-conventions==0.60b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==25.0
    # via opentelemetry-instrumentation
pathable==0.4.4
    # via jsonschema-path
pathvalidate==3.3.1
    # via py-key-value-aio
pillow==11.3.0
    # via strands-agents-tools
platformdirs==4.5.1
    # via
    #   fastmcp
    #   virtualenv
pre-commit==4.5.0
    # via bedrock-agentcore
prompt-toolkit==3.0.52
    # via
    #   strands-agents-builder
    #   strands-agents-tools
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
py-key-value-aio==0.3.0
    # via fastmcp
py-key-value-shared==0.3.0
    # via py-key-value-aio
pycparser==2.23
    # via cffi
pycryptodome==3.23.0
    # via teradatasql
pydantic==2.12.5
    # via
    #   bedrock-agentcore
    #   fastmcp
    #   mcp
    #   ollama
    #   openapi-pydantic
    #   pydantic-settings
    #   strands-agents
    #   teradata-mcp-server
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via mcp
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via
    #   mcp
    #   strands-agents-tools
pyperclip==1.11.0
    # via fastmcp
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.2.1
    # via
    #   fastmcp
    #   mcp
    #   pydantic-settings
    #   teradata-mcp-server
python-multipart==0.0.20
    # via mcp
pywin32==311
    # via mcp
pyyaml==6.0.3
    # via
    #   jsonschema-path
    #   pre-commit
    #   teradata-mcp-server
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   aws-requests-auth
    #   jsonschema-path
    #   strands-agents-tools
rich==14.2.0
    # via
    #   cyclopts
    #   fastmcp
    #   rich-rst
    #   strands-agents-builder
    #   strands-agents-tools
    #   typer
rich-rst==1.3.2
    # via cyclopts
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.16.0
    # via boto3
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   halo
    #   markdownify
    #   python-dateutil
slack-bolt==1.27.0
    # via strands-agents-tools
slack-sdk==3.39.0
    # via slack-bolt
soupsieve==2.8
    # via beautifulsoup4
spinners==0.0.24
    # via halo
sqlalchemy==2.0.45
    # via
    #   teradata-mcp-server
    #   teradatasqlalchemy
sse-starlette==3.0.3
    # via mcp
starlette==0.50.0
    # via
    #   bedrock-agentcore
    #   mcp
strands-agents==1.19.0
    # via
    #   tdmcpagentcore (pyproject.toml)
    #   strands-agents-builder
    #   strands-agents-tools
strands-agents-builder==0.1.10
    # via tdmcpagentcore (pyproject.toml)
strands-agents-tools==0.2.17
    # via
    #   tdmcpagentcore (pyproject.toml)
    #   strands-agents-builder
sympy==1.14.0
    # via strands-agents-tools
tenacity==9.1.2
    # via strands-agents-tools
teradata-mcp-server==0.1.6
    # via tdmcpagentcore (pyproject.toml)
teradatasql==20.0.0.47
    # via teradatasqlalchemy
teradatasqlalchemy==20.0.0.8
    # via teradata-mcp-server
termcolor==3.2.0
    # via halo
typer==0.20.0
    # via mcp
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   beautifulsoup4
    #   bedrock-agentcore
    #   exceptiongroup
    #   mcp
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   py-key-value-shared
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   starlette
    #   strands-agents
    #   strands-agents-tools
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via strands-agents-tools
urllib3==2.6.1
    # via
    #   bedrock-agentcore
    #   botocore
    #   requests
uvicorn==0.38.0
    # via
    #   bedrock-agentcore
    #   fastmcp
    #   mcp
virtualenv==20.35.4
    # via pre-commit
watchdog==6.0.0
    # via
    #   strands-agents
    #   strands-agents-tools
wcwidth==0.2.14
    # via prompt-toolkit
websockets==15.0.1
    # via
    #   bedrock-agentcore
    #   fastmcp
wrapt==1.17.3
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-threading
yarl==1.22.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
